% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{backprop}
\alias{backprop}
\title{Backpropagation}
\usage{
backprop(model, error, forward_pass)
}
\arguments{
\item{model}{list of all the weights and biases}

\item{error}{gradients to the output of the network}

\item{forward_pass}{intermediate values from the forward pass}
}
\value{
list of derivatives after the backwards pass
}
\description{
Runs a backwards pass through the network.
}
