% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-train.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Model}
\usage{
evaluate(inputs, target, model)
}
\arguments{
\item{inputs}{set of inputs to the model}

\item{target}{set of targets in one-hot encoded form}

\item{model}{list of weights and biases}
}
\value{
accuracy of the model
}
\description{
Evaluates the performance of a model on a given dataset.
}
\examples{
\dontrun{
evaluate(train_data, train_target, mlp_model)
}
}
