% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{forwardprop}
\alias{forwardprop}
\title{Forward propagation}
\usage{
forwardprop(model, x)
}
\arguments{
\item{model}{list of all the weights and biases}

\item{x}{input to the network}
}
\value{
list of all intermediate values
}
\description{
Runs a forward pass through the network.
}
