% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{init_nn}
\alias{init_nn}
\title{Initialize network}
\usage{
init_nn(num_inputs, num_hidden_1, num_hidden_2, num_outputs)
}
\arguments{
\item{num_inputs}{dimension of inputs}

\item{num_hidden_1}{dimension of first hidden layer}

\item{num_hidden_2}{dimension of second hidden layer}

\item{num_outputs}{dimension of output}
}
\value{
list containing weight and bias matrices in each layer of the network
}
\description{
Initialize 3 layer fully connected neural network, also known as multilayer
perceptron, setting biases to 0 and using the Xavier initialization method
for weights.
}
\examples{
mlp_model <- init_nn(784, 100, 50, 10)
}
