% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-mnist.R
\name{load_mnist_test}
\alias{load_mnist_test}
\title{Load Testing Data}
\usage{
load_mnist_test()
}
\value{
list of testing data and targets
}
\description{
Loads MNIST testing data and generates one hot encodings for the targets.
}
\examples{
\donttest{
mnist_test <- load_mnist_test()
test_data <- mnist_test[1]
test_target <- mnist_test[2]
}
}
