% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-mnist.R
\name{load_mnist_train}
\alias{load_mnist_train}
\title{Load Training Data}
\usage{
load_mnist_train(train_prop = 0.7)
}
\arguments{
\item{train_prop}{the proportion of the data used for training data; the
remaining will be used for validation}
}
\value{
list of training and validation data and targets
}
\description{
Loads MNIST training and validation data and generates one hot encodings
for the targets.
}
\examples{
\donttest{
mnist_train <- load_mnist_train()
train_data <- mnist_train[1]
train_target <- mnist_train[2]
validate_data <- mnist_train[3]
validate_target <- mnist_train[4]
}
}
