% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-mnist.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{One Hot Encoding}
\usage{
one_hot_encoding(data_raw, ncat = 10)
}
\arguments{
\item{data_raw}{data input to create encoding; target must be first column}

\item{ncat}{number of categories to use for the encoding}
}
\value{
targets in a one hot encoding matrix
}
\description{
Creates a one hot encoding matrix with the specified number of categories
for the targets. Target must be the first column of the data_raw input.
}
