% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-plot.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{Plot Accuracy}
\usage{
plot_accuracy(accuracy_train, accuracy_validate)
}
\arguments{
\item{accuracy_train}{list of training accuracy}

\item{accuracy_validate}{list of validation accuracy}
}
\description{
Plot the training and validation accuracy.
}
