% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{update}
\alias{update}
\title{Update Model}
\usage{
update(model, back_pass, alpha)
}
\arguments{
\item{model}{list of all the weights and biases}

\item{back_pass}{derivatives from a backwards pass through the network}

\item{alpha}{learning rate}
}
\value{
updated list of the weights and biases
}
\description{
Updates the model using derivatives from a backward pass.
}
