% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-train.R
\name{train_nn}
\alias{train_nn}
\title{Train Network}
\usage{
train_nn(
  train_set,
  target_set,
  model,
  alpha,
  epochs,
  batch_size = nrow(train_set)
)
}
\arguments{
\item{train_set}{set of training data}

\item{target_set}{set of targets in one-hot encoded form}

\item{model}{list of weights and biases}

\item{alpha}{learning rate}

\item{epochs}{number of epochs}

\item{batch_size}{mini-batch size}
}
\value{
list of weights and biases after training
}
\description{
Train the network with specified hyperparameters.
}
\examples{
train_nn(train, target, model, 0.01, 100, 25)
}
